+++ src/main/java/com/ahumadamob/fnanz/service/UsuarioService.java
@@ -1,6 +1,9 @@
 package com.ahumadamob.fnanz.service;
 
-import com.ahumadamob.fnanz.dto.*;
+import com.ahumadamob.fnanz.dto.UsuarioCreateDto;
+import com.ahumadamob.fnanz.dto.UsuarioPatchDto;
+import com.ahumadamob.fnanz.dto.UsuarioPasswordDto;
+import com.ahumadamob.fnanz.dto.response.UsuarioResponseDto;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
@@ -8,12 +11,12 @@ import org.springframework.data.domain.Pageable;
  * Servicio para operaciones de {@link com.ahumadamob.fnanz.entity.Usuario}.
  */
 public interface UsuarioService {
-    UsuarioDto create(UsuarioCreateDto dto);
-    Page<UsuarioDto> list(String q, Boolean activo, Pageable pageable);
-    UsuarioDto get(Long id);
-    UsuarioDto update(Long id, UsuarioPatchDto dto);
+    UsuarioResponseDto create(UsuarioCreateDto dto);
+    Page<UsuarioResponseDto> list(String q, Boolean activo, Pageable pageable);
+    UsuarioResponseDto get(Long id);
+    UsuarioResponseDto update(Long id, UsuarioPatchDto dto);
     void delete(Long id);
     void restore(Long id);
     void changePassword(Long id, UsuarioPasswordDto dto);
-    UsuarioDto me();
+    UsuarioResponseDto me();
 }
